fix(design-system): Lazy-load node icon SVGs to reduce main bundle size
Split nodeIconSet into a separate node-icons.ts module that is
dynamically imported only when a node: prefixed icon is first rendered.
This moves ~80 node SVGs out of the base bundle into an async chunk.
